# application/rest-api-adapter/src/main/resources/openapi-contract.yaml
openapi: 3.0.3
info:
  title: Example API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local server
  - url: https://staging-api.gebio.app
    description: Staging server

paths:
  /health/check:
    get:
      summary: Check health of the API
      operationId: check
      tags:
        - Health
      responses:
        "200":
          description: Application is running
          content:
            application/json:
              schema:
                type: string
                example: "OK, API is running"
  /api/v1/me:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get current logged in user
      description: Get current logged in user
      operationId: getCurrentUser
      tags:
        - User
      responses:
        "200":
          description: Successfully retrieved current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponseContract'
              example:
                user:
                  id: 77dbf8f5-8490-4986-9863-969bf7dc6c23
                  email: john.doe@gmail.com
        "401":
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/user/{user_id}/board:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get all the retrospective boards for a given user id
      description: Get all the retrospective boards for a given user id
      operationId: findBoards
      tags:
        - User
      parameters:
        - name: user_id
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successfully retrieved boards for user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindBoardsResponseContract'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/public/guest:
    post:
      security:
        - bearerAuth: [ ]
      summary: Create guest user
      description: Create guest user
      operationId: createGuest
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGuestRequestContract'
      responses:
        "200":
          description: Successfully created guest user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGuestResponseContract'
  /api/v1/public/board/{board_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Find retrospective board by its id
      description: Find retrospective board by its id
      operationId: findById
      tags:
        - Board
      parameters:
        - name: board_id
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successfully retrieved retrospective board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindBoardResponseContract'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          description: Board was not found
  /api/v1/board:
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a retrospective board
      description: Create a retrospective board
      operationId: createBoard
      tags:
        - Board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardRequestContract'
      responses:
        "201":
          description: Successfully created retrospective board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBoardResponseContract'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/board/{board_id}:
    patch:
      security:
        - bearerAuth: [ ]
      summary: Update board
      description: Allow authenticated user to update board
      operationId: updateBoard
      tags:
        - Board
      parameters:
        - name: board_id
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBoardRequestContract'
      responses:
        "200":
          description: Successfully renamed board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateBoardResponseContract'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/ForbiddenError'
        "404":
          description: Board was not found
  /api/v1/public/board/{board_id}/add-user:
    post:
      security:
        - bearerAuth: [ ]
      summary: Add user to a board
      description: Add user to a board
      operationId: addUserToBoard
      tags:
        - Board
      parameters:
        - name: board_id
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserToBoardRequestContract'
      responses:
        "200":
          description: Successfully added user to retrospective board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddUserToBoardResponseContract'
        "404":
          description: Board was not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AddUserToBoardRequestContract:
      description: Add current user to retrospective board request contract
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          format: uuid
    AddUserToBoardResponseContract:
      description: Add current user to retrospective board response contract
      type: object
      required:
        - board
      properties:
        board:
          $ref: '#/components/schemas/BoardContract'
    UpdateBoardRequestContract:
      description: Update retrospective board request contract
      type: object
      required:
        - boardName
      properties:
        boardName:
          type: string
          minLength: 1
    UpdateBoardResponseContract:
      description: Update retrospective board response contract
      type: object
      required:
        - board
      properties:
        board:
          $ref: '#/components/schemas/BoardContract'
    CreateGuestRequestContract:
      description: Create guest user request contract
      type: object
      required:
        - username
      properties:
        username:
          type: string
    CreateGuestResponseContract:
      description: Create guest user response contract
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/UserContract'
    FindBoardResponseContract:
      description: Find retrospective board response contract
      type: object
      properties:
        board:
          $ref: '#/components/schemas/BoardContract'
    CreateBoardResponseContract:
      description: Retrospective board creation response contract
      type: object
      properties:
        board:
          $ref: '#/components/schemas/BoardContract'
    CreateBoardRequestContract:
      description: Retrospective board creation request contract
      type: object
      required:
        - title
        - templateId
      properties:
        title:
          type: string
        templateId:
          type: string
          format: uuid
    CurrentUserResponseContract:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/UserContract'
    BoardContract:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        templateId:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        cards:
          type: array
          items:
            $ref: '#/components/schemas/CardContract'
        members:
          type: array
          items:
            $ref: '#/components/schemas/UserContract'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CardContract:
      type: object
      required:
        - id
        - content
        - color
        - position
        - ownerInfo
        - votes
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        color:
          type: string
        position:
          $ref: '#/components/schemas/CardPositionContract'
        ownerInfo:
          $ref: '#/components/schemas/CardOwnerInfoContract'
        boardId:
          type: string
          format: uuid
        votes:
          type: integer
    CardPositionContract:
      type: object
      required:
        - posX
        - posY
      properties:
        posX:
          type: integer
        posY:
          type: integer
    CardOwnerInfoContract:
      type: object
      required:
        - id
        - email
        - logo
        - role
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        logo:
          type: string
        username:
          type: string
        role:
          $ref: '#/components/schemas/UserRoleContract'
    UserContract:
      type: object
      required:
        - id
        - email
        - logo
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        logo:
          type: string
        username:
          type: string
        role:
          $ref: '#/components/schemas/UserRoleContract'
    AddCardRequestContract:
      type: object
      properties:
        boardId:
          type: string
          format: uuid
        card:
          $ref: '#/components/schemas/CardContract'
    AddCardResponseContract:
      type: object
      properties:
        card:
          $ref: '#/components/schemas/CardContract'
        operationType:
          $ref: '#/components/schemas/OperationTypeContract'
    UpdateCardRequestContract:
      type: object
      properties:
        card:
          $ref: '#/components/schemas/CardContract'
    UpdateCardResponseContract:
      type: object
      properties:
        card:
          $ref: '#/components/schemas/CardContract'
        operationType:
          $ref: '#/components/schemas/OperationTypeContract'
    DeleteCardRequestContract:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
    DeleteCardResponseContract:
      type: object
      properties:
        card:
          $ref: '#/components/schemas/CardContract'
        operationType:
          $ref: '#/components/schemas/OperationTypeContract'
    FindBoardsResponseContract:
      type: object
      properties:
        boards:
          type: array
          items:
            $ref: '#/components/schemas/BoardContract'
    OperationTypeContract:
      type: string
      enum:
        - ADD
        - UPDATE
        - DELETE
    UserRoleContract:
      type: string
      enum:
        - USER
        - GUEST
  responses:
    UnauthorizedError:
      description: Authentication is missing or invalid
    ForbiddenError:
      description: User is not allowed to perform this action